Move model into repository implementation
