+++ packages/backend/src/app/Users/UserRepository.ts
@@ -1,32 +1,4 @@
-import {
-  Column,
-  CreateDateColumn,
-  DeleteDateColumn,
-  Entity,
-  Index,
-  PrimaryGeneratedColumn,
-  UpdateDateColumn,
-} from 'typeorm';
 import User from './User';
-
-@Entity({ name: 'user' })
-export class UserModel {
-  @PrimaryGeneratedColumn()
-  id!: number;
-
-  @Column({ type: 'text', unique: true })
-  @Index({ unique: true })
-  name!: string;
-
-  @CreateDateColumn()
-  created!: Date;
-
-  @UpdateDateColumn()
-  updated?: Date;
-
-  @DeleteDateColumn()
-  deleted?: Date;
-}
 
 export default interface IUserRepository {
   findById(id: number): Promise<User>;
